requirements.txt
# Core Dependencies
dash==2.14.1
dash-bootstrap-components>=1.4.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
plotly==5.18.0
python-dotenv==1.0.0
pyyaml==6.0.1
numpy==1.26.2
pandas==2.1.3

# Data Processing
scipy>=1.11.0

# Database Connections
influxdb-client>=1.36.0
pyodbc>=4.0.39
SQLAlchemy>=2.0.0
psycopg2-binary==2.9.9

# AI and Analysis
requests>=2.31.0
jsonschema>=4.17.3
python-dateutil>=2.8.2

# Excel Support
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Logging and Utilities
coloredlogs>=15.0.1
humanfriendly>=10.0

# Testing and Development
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
black>=23.11.0
flake8>=6.1.0
